fix: hard-limit kMaxVars to 255 via static_assert; default kMaxChannels=8 with CYMON_MAX_CHANNELS override
Agent-Logs-Url: https://github.com/Tecnologic/cymon-lib/sessions/801a3240-6068-428b-a551-5bbd5212769b

+++ src/cymon/device.cpp
@@ -24,7 +24,7 @@ Device::Device() = default;
 void Device::set_actual_sample_period_us(float us) { actual_sample_period_us_ = us; }
 
 bool Device::RegisterVariable(std::string_view name, std::string_view unit, std::function<float()> getter) {
-  if (var_count_ >= static_cast<uint8_t>(kMaxVars)) {
+  if (var_count_ >= kMaxVars) {
     return false;
   }
   if (name.size() > 32U) {
